add Pundit and trackable to Devise

+++ app/controllers/companies_controller.rb
@@ -2,27 +2,11 @@ class CompaniesController < ApplicationController
   before_action :authenticate_user!
 
 
-#  def update_cities
-#    @cities = City.where("country_id = ?", params[:country_id])
-#    respond_to do |format|
-#      format.js
-#    end
-#  end
-
-  def select_his
-    #filter using id passed by ajax
-    @selected_history = CompanyHistory.find_by(id: params[:send_id])
- 
-    render json: @selected_history
-    #respond_to do |format|
-    #  format.json { render json: @selected_history }
-    #end
-
-  end 
-
   # GET /companies
   # GET /companies.json
   def index
+    authorize(Company)
+    #@companies = policy_scope(Company)
     respond_to do |format|
       format.html
     end   
@@ -30,8 +14,9 @@ class CompaniesController < ApplicationController
 
   # POST /companies
   def datatables_index
+    data_scope = current_user.admin? ? -1 : current_user.id
     respond_to do |format|
-      format.json{ render json: CompanyDatatable.new(view_context, { only_for_current_user_id: current_user.id }) }
+      format.json{ render json: CompanyDatatable.new(view_context, { only_for_current_user_id: data_scope }) }
     end
   end
 
@@ -39,6 +24,8 @@ class CompaniesController < ApplicationController
   # GET /companies/1.json
   def show
     @company = load_company
+
+    authorize @company
 
     respond_to do |format|
       format.html
@@ -51,6 +38,8 @@ class CompaniesController < ApplicationController
     @company = Company.new
     @company.user = current_user
 
+    authorize @company
+
     respond_to do |format|
       format.html
     end
@@ -59,6 +48,8 @@ class CompaniesController < ApplicationController
   # GET /companies/1/edit
   def edit
     @company = load_company
+
+    authorize @company
 
     respond_to do |format|
       format.html
@@ -70,6 +61,8 @@ class CompaniesController < ApplicationController
   def create
     @company = Company.new(company_params)
     @company.user = current_user
+
+    authorize @company
 
     respond_to do |format|
       if @company.save
@@ -87,6 +80,8 @@ class CompaniesController < ApplicationController
   def update
     @company = load_company
         
+    authorize @company
+
     respond_to do |format|
       if @company.update(company_params)
         format.html { redirect_to @company, success: t('activerecord.messages.successfull.updated', data: @company.short) }
@@ -103,6 +98,8 @@ class CompaniesController < ApplicationController
   def destroy
     @company = load_company    
 
+    authorize @company
+
     if @company.destroy
       redirect_to companies_url, success: t('activerecord.messages.successfull.destroyed', data: @company.short)
     else 
@@ -114,7 +111,11 @@ class CompaniesController < ApplicationController
   private
 
     def load_company
-      Company.by_user(current_user.id).find(params[:id])
+      if current_user.admin?
+        Company.find(params[:id])
+      else
+        Company.by_user(current_user.id).find(params[:id])
+      end
     end
 
     # Never trust parameters from the scary internet, only allow the white list through.
